Unione modifiche. Aggiunto helper per db per le Room, portato avanti lavoro su indoor engine, da adesso si lavora sul nuovo server (nuovo DB)

+++ app/src/main/java/micc/beaconav/indoorEngine/spot/Spot.java
@@ -27,7 +27,7 @@ public class Spot implements ISpot
     public Spot(PointF position, int floor){
         _indoorPosition = new IndoorPosition(position, floor);
     }
-    public Spot(int x, int y, int floor){
+    public Spot(float x, float y, int floor){
         _indoorPosition = new IndoorPosition(x, y, floor);
     }
